# Avoid ToCharArray allocation in JsonCamelCaseNamingPolicy