Microbenchmark calling ConvertName with either "PascalCase" or "camelCased":
```
| Method |       Text |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |----------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|    Old | PascalCase | 35.724 ns | 1.3263 ns | 3.7410 ns | 0.0458 |     - |     - |      96 B |
|    New | PascalCase | 26.932 ns | 0.7734 ns | 2.2683 ns | 0.0229 |     - |     - |      48 B |
|    Old | camelCased |  3.323 ns | 0.1484 ns | 0.4211 ns |      - |     - |     - |         - |
|    New | camelCased |  2.685 ns | 0.1435 ns | 0.4186 ns |      - |     - |     - |         - |
```
cc: @steveharter, @ahsonkhan